+++ store/src/test/java/com/nytimes/android/external/store4/impl/FlowStoreTest.kt
@@ -1,4 +1,4 @@
-package com.nytimes.android.external.store3.pipeline
+package com.nytimes.android.external.store4.impl
 
 import com.nytimes.android.external.store4.*
 import com.nytimes.android.external.store4.ResponseOrigin.Cache
